New Quiz features
A timer is added to the top of the quiz. Exams are auto submitted when the time ends. Student Send request to the admin for sigining up. Fixing some errors. Fix student score calculation. Edit options of the question. Considering the senario when the student does not answer some questions